+++ src/connectors/__test__/commentService.test.ts
@@ -1,7 +1,4 @@
 import { CommentService } from '../commentService'
-import { knex } from 'connectors/db'
-
-afterAll(knex.destroy)
 
 const commentService = new CommentService()