+++ src/types/__test__/article.test.ts
@@ -2,7 +2,6 @@ import _ from 'lodash'
 // internal
 import { toGlobalId } from 'common/utils'
 import { PUBLISH_STATE } from 'common/enums'
-import { knex } from 'connectors/db'
 import {
   GQLNodeInput,
   GQLPublishArticleInput,
@@ -10,8 +9,6 @@ import {
 } from 'definitions'
 // local
 import { testClient, publishArticle, putDraft, getViewerMAT } from './utils'
-
-afterAll(knex.destroy)
 
 const mediaHash = 'someIpfsMediaHash1'
 
@@ -130,7 +127,7 @@ export const appreciateArticle = async (input: GQLAppreciateArticleInput) => {
   return article
 }
 
-describe('query article', async () => {
+describe('query article', () => {
   test('query articles', async () => {
     const { query } = await testClient({ isAuth: true, isAdmin: true })
     const result = await query({
@@ -152,7 +149,7 @@ describe('query article', async () => {
   })
 })
 
-describe('query tag on article', async () => {
+describe('query tag on article', () => {
   test('query tag on article', async () => {
     const id = toGlobalId({ type: 'Article', id: 1 })
     const { query } = await testClient()
@@ -168,7 +165,7 @@ describe('query tag on article', async () => {
   })
 })
 
-describe('publish article', async () => {
+describe('publish article', () => {
   test('create draft, publish and recall', async () => {
     jest.setTimeout(10000)
     const draft = {
@@ -230,7 +227,7 @@ describe('publish article', async () => {
   })
 })
 
-describe('appreciate article', async () => {
+describe('appreciate article', () => {
   test('appreciate success', async () => {
     const viewerCurrentMAT = await getViewerMAT()
     const articleCurrentMAT = await getArticleMAT({ id: ARTICLE_ID })
@@ -255,7 +252,7 @@ describe('appreciate article', async () => {
   })
 })
 
-describe('report article', async () => {
+describe('report article', () => {
   test('report a article without assets', async () => {
     const { mutate } = await testClient({ isAuth: true })
     const result = await mutate({
@@ -290,7 +287,7 @@ describe('report article', async () => {
   })
 })
 
-describe('toggle article state', async () => {
+describe('toggle article state', () => {
   test('enable article live', async () => {
     const { mutate } = await testClient({ isAuth: true, isAdmin: true })
     const result = await mutate({