+++ src/types/__test__/user.test.ts
@@ -8,7 +8,7 @@ import {
 } from 'common/enums'
 import { MaterializedView } from 'definitions'
 import { UserService } from 'connectors'
-import { knex, refreshView } from 'connectors/db'
+import { refreshView } from 'connectors/db'
 import {
   defaultTestUser,
   getUserContext,
@@ -22,7 +22,6 @@ beforeAll(async () => {
   userService = new UserService()
   // await userService.initSearch()
 })
-afterAll(knex.destroy)
 
 const USER_LOGIN = `
   mutation UserLogin($input: UserLoginInput!) {
@@ -375,7 +374,7 @@ describe('register and login functionarlities', () => {
   })
 })
 
-describe('user mat', async () => {
+describe('user mat', () => {
   test('total', async () => {
     const mat = await getViewerMAT()
     expect(typeof mat).toBe('number')
@@ -626,7 +625,7 @@ describe('user recommendations', () => {
   })
 })
 
-describe('badges', async () => {
+describe('badges', () => {
   test('get user badges', async () => {
     const { query } = await testClient({
       isAuth: true
@@ -640,7 +639,7 @@ describe('badges', async () => {
   })
 })
 
-describe('verification code', async () => {
+describe('verification code', () => {
   const email = `verification-${Math.floor(Math.random() * 100)}@test.com`
   const type = 'register'