+++ src/types/system.ts
@@ -8,11 +8,16 @@ export default /* GraphQL */ `
   }
 
   extend type Mutation {
+    "Upload a single file."
     singleFileUpload(input: SingleFileUploadInput!): Asset! @authenticate
+
+    "Delete a uploaded file."
     singleFileDelete(input: SingleFileDeleteInput!): Boolean! @authenticate
     feedback(input: FeedbackInput!): Boolean
     setBoost(input: SetBoostInput!): Node! @authorize
     putRemark(input: PutRemarkInput!): String @authorize
+
+    "Add specific user behavior record."
     logRecord(input: LogRecordInput!): Boolean
   }
 
@@ -36,13 +41,15 @@ export default /* GraphQL */ `
     pageInfo: PageInfo!
   }
 
+  "This type contains system-wise settings."
   type Official {
     reportCategory: [Category!]!
     feedbackCategory: [Category!]!
     releases(input: ReleasesInput!): [Release!]
+
+    "Links of specific pages on Matters site."
     links: OfficialLinks!
     placements: Placements!
-    gatewayUrls: [URL!]
   }
 
   type OSS {
@@ -95,10 +102,20 @@ export default /* GraphQL */ `
     adLabel: Boolean!
   }
 
+  """
+  This type contains type, link and related data of an asset.
+  """
   type Asset {
+    "Unique ID of this Asset."
     id: ID!
+
+    "Types of this asset."
     type: AssetType!
+
+    "Link of this asset."
     path: String!
+
+    "Time of this asset was created."
     createdAt: DateTime!
   }
 
@@ -248,6 +265,7 @@ export default /* GraphQL */ `
     ReadResponseInfoPopUp
   }
 
+  "Enums for asset types."
   enum AssetType {
     avatar
     cover
@@ -274,6 +292,7 @@ export default /* GraphQL */ `
     googlePlay
   }
 
+  "Enums for user roles."
   enum Role {
     vistor
     user
@@ -281,13 +300,13 @@ export default /* GraphQL */ `
   }
 
   input CostComplexity {
-    min: Int = 1,
+    min: Int = 1
     max: Int
   }
 
   directive @cost(
-    multipliers: [String],
-    useMultipliers: Boolean,
+    multipliers: [String]
+    useMultipliers: Boolean
     complexity: CostComplexity
   ) on OBJECT | FIELD_DEFINITION
 
@@ -295,13 +314,9 @@ export default /* GraphQL */ `
     reason: String = "No longer supported"
   ) on FIELD_DEFINITION | ENUM_VALUE
 
-  directive @authenticate(
-    requires: Role = user,
-  ) on OBJECT | FIELD_DEFINITION
+  directive @authenticate(requires: Role = user) on OBJECT | FIELD_DEFINITION
 
-  directive @authorize(
-    requires: Role = admin,
-  ) on OBJECT | FIELD_DEFINITION
+  directive @authorize(requires: Role = admin) on OBJECT | FIELD_DEFINITION
 
   directive @private on FIELD_DEFINITION
 `